+++ src/main/java/view/gamer/DealerOutputView.java
@@ -1,0 +1,29 @@
+package view.gamer;
+
+import domain.card.Card;
+import dto.DealerDTO;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class DealerOutputView {
+    public static void print(DealerDTO dealerDTO) {
+        String outputWithoutSummationCardPoint = generateOutputWithoutSummationCardPoint(dealerDTO);
+        String summationCardPointOutput = "결과: %d".formatted(dealerDTO.getSummationCardPoint());
+        System.out.printf("%s - %s%n", outputWithoutSummationCardPoint, summationCardPointOutput);
+    }
+
+    private static String generateOutputWithoutSummationCardPoint(DealerDTO dealerDTO) {
+        String name = "딜러";
+        List<Card> cards = dealerDTO.getHoldingCards();
+        String nameOutput = name + "카드";
+        String cardsOutput = cards.stream()
+                .map(CardOutputGenerator::generate)
+                .collect(Collectors.joining(", "));
+        return "%s: %s".formatted(nameOutput, cardsOutput);
+    }
+
+    public static void printWithoutSummationCardPoint(DealerDTO dealerDTO) {
+        String outputWithoutSummationCardPoint = generateOutputWithoutSummationCardPoint(dealerDTO);
+        System.out.println(outputWithoutSummationCardPoint);
+    }
+}
